requirements.txt
typing-extensions==3.7.4.2
ecdsa==0.16.0
pysha3==1.0.2
eth-keys==0.3.3
rlp
web3==5.22.0
solana==0.10.0
psycopg2-binary
ethereum
py-solc-x==1.1.0
flask==2.0.1
prometheus_client==0.13.1
git+https://github.com/neonlabsorg/python-logged-groups.git@2.1.5
